refactor: break down monolithic discount component for better maintainability
- Split 679-line component into focused components and custom hook
- Add centralized formatting utilities
- Improve code organization and readability
